* client-src/client_util.c: Fix compiler error
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@6363 a8d146d6-cc15-0410-8900-af154a0219e0

+++ client-src/client_util.c
@@ -1788,8 +1788,8 @@ check_access_message(
     char  ruid_str[NUM_STR_SIZE];
     char  euid_str[NUM_STR_SIZE];
 
-    g_snprintf(ruid_str, sizeof(ruid_str), "%d", getuid());
-    g_snprintf(euid_str, sizeof(euid_str), "%d", geteuid());
+    g_snprintf(ruid_str, sizeof(ruid_str), "%d", (int)getuid());
+    g_snprintf(euid_str, sizeof(euid_str), "%d", (int)geteuid());
 
     if(mode == F_OK)
         noun = "find", adjective = "exists";
